Working on a UTF8 version of the NEXI parser.
git-svn-id: https://groke.otago.ac.nz/ant/ant@431 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/stemmer_term_similarity.c
@@ -19,7 +19,7 @@
 */
 ANT_stemmer_term_similarity::ANT_stemmer_term_similarity(ANT_search_engine *search_engine, ANT_stemmer *stemmer) : ANT_stemmer(search_engine)
 {
-this->buffer = new long [search_engine->document_count()];
+this->buffer = new long [(size_t)(search_engine->document_count())];
 this->search_engine = search_engine;
 this->base_stemmer = stemmer;
 }
